+++ SwiftyRSA/PublicKey.swift
@@ -94,14 +94,9 @@ public class PublicKey: Key {
             range: all
         )
         
-        let keys = matches.flatMap { result -> PublicKey? in
+        let keys = matches.compactMap { result -> PublicKey? in
             
-            #if swift(>=4.0)
             let match = result.range(at: 1)
-            #else
-            let match = result.rangeAt(1)
-            #endif
-            
             let start = pemString.index(pemString.startIndex, offsetBy: match.location)
             let end = pemString.index(start, offsetBy: match.length)